Rename github-database-prompt to github-database-download
