+++ extensions/ql-vscode/src/databases/github-database-download.ts
